Remove Un-documented Filtering Feature
Using a filter to filter out items during pagination would produce unpredictable results such as fewer items per page than the current page size.

Using a filter to iterate over and mutate individual items is a questionable feature at best, and, something highly project-specific for which a dedicated adapter would be a better fit.

None of this is documented and there is no clear use-case for the feature.

Removing it also means we can remove the dependency on `laminas-filter`

Signed-off-by: George Steel <george@net-glue.co.uk>