Remove the Scrolling Style Plugin Manager, reduce static properties
Whilst it is possible to create custom scrolling styles, it is unlikely that many users do.

Furthermore, the reliance on static properties for the default style make for state that is difficult to reason about and to test.

This patch removes the plugin manager because:

- Scrolling styles are simple computations and all shipped implementations are "newable"
- Bringing in the baggage of a plugin manager for these simple objects is overkill

Internally, a static factory (marked as `@internal`) is used to `new` the relevant scrolling style.

The style can be overridden by passing a string or instance to the paginators constructor, or, by passing the same thing to the `getPages()` method which is the only place scrolling styles are used.

This patch does not address the concept of project-wide defaults. This will be addressed in a later patch that will introduce a `PaginatorFactory` - a service that will accept configured defaults and be used to generate new paginators, much like the existing `Factory` class.

Apart from the plugin manager removal, the `Paginator` class has the following (static) methods removed:

- `getDefaultScrollingStyle`
- `setDefaultScrollingStyle`
- `getDefaultItemCountPerPage`
- `setDefaultItemCountPerPage`
- `setScrollingStylePluginManager`
- `getScrollingStylePluginManager`

Signed-off-by: George Steel <george@net-glue.co.uk>